+++ packages/editor/editor-core/src/__tests__/unit/plugins/insert-block/ui/ToolbarInsertBlock.tsx
@@ -50,8 +50,8 @@ import Icon from '@atlaskit/icon';
 const emojiProvider = emojiData.testData.getEmojiResourcePromise();
 
 const mediaProvider: Promise<MediaProvider> = Promise.resolve({
-  viewContext: Promise.resolve({} as any),
-  uploadContext: Promise.resolve({} as any),
+  viewMediaClientConfig: {} as any,
+  uploadMediaClientConfig: {} as any,
 });
 
 const providerFactory = ProviderFactory.create({ mediaProvider });