+++ packages/editor/editor-core/src/plugins/type-ahead/commands/select-item.ts
@@ -136,7 +136,7 @@ export const selectItem = (
 
         tr = tr.replaceWith(start, start, fragment);
 
-        // This problem affects Chrome v58-62. See: https://github.com/ProseMirror/prosemirror/issues/710
+        // This problem affects Chrome v58+. See: https://github.com/ProseMirror/prosemirror/issues/710
         if (isChromeWithSelectionBug) {
           const selection = document.getSelection();
           if (selection) {