+++ packages/editor/email-renderer/src/nodes/table-header.ts
@@ -1,26 +1,28 @@
 import { N20, N50 } from '@atlaskit/adf-schema';
-import { createTag, serializeStyle } from '../util';
+import { createTag } from '../create-tag';
+import { serializeStyle } from '../serialize-style';
+import { createClassName } from '../styles/util';
 import { NodeSerializerOpts } from '../interfaces';
 
-const baseStyle = {
-  'background-color': N20,
-  'background-clip': 'padding-box',
-  border: `1px solid ${N50}`,
-  'border-right-width': 0,
-  'border-bottom-width': 0,
-  'font-weight': 'bold',
-  height: '2.5em',
-  'min-width': '3em',
-  padding: '6px 10px',
-  'text-align': 'left',
-  'vertical-align': 'top',
-};
+export const styles = `
+.${createClassName('tableHeader')} {
+  background-clip: padding-box;
+  border: 1px solid ${N50};
+  border-right-width: 0;
+  border-bottom-width: 0;
+  font-weight: bold;
+  height: 2.5em;
+  min-width: 3em;
+  padding: 6px 10px;
+  text-align: left;
+  vertical-align: top;
+}
+`;
 
 export default function tableHeader({ attrs, text }: NodeSerializerOpts) {
   const { colspan, rowspan, background } = attrs;
   const style = serializeStyle({
-    ...baseStyle,
-    'background-color': background ? background : baseStyle['background-color'],
+    'background-color': background ? background : N20,
   });
 
   return createTag(
@@ -29,6 +31,7 @@ export default function tableHeader({ attrs, text }: NodeSerializerOpts) {
       colspan,
       rowspan,
       style,
+      class: createClassName('tableHeader'),
     },
     text,
   );