+++ packages/editor/email-renderer/src/styles/util.ts
@@ -1,0 +1,4 @@
+const PREFIX = 'pfcs-generated';
+export const createClassName = (name: string) => {
+  return `${PREFIX}-${name}`;
+};