+++ packages/search/global-search/example-helpers/withNavigation.tsx
@@ -24,6 +24,7 @@ const Radio = styled.input`
 
 export interface Config {
   hideLocale?: boolean;
+  context?: 'jira' | 'confluence';
   message?: JSX.Element;
   cloudIds?: {
     [k: string]: string;
@@ -41,7 +42,7 @@ const MessageContainer = styled.div`
 `;
 
 interface State {
-  context: 'home' | 'jira' | 'confluence';
+  context: 'jira' | 'confluence';
   locale: string;
 }
 
@@ -49,28 +50,32 @@ interface State {
 export default function withNavigation<P extends Props>(
   WrappedComponent: ComponentType<P>,
   props?: Config,
-  availableContext: QuickSearchContext[] = ['confluence', 'jira', 'home'],
+  availableContext: QuickSearchContext[] = ['confluence', 'jira'],
   drawerIsOpen?: boolean,
 ): ComponentType<Partial<P>> {
-  return class WithNavigation extends React.Component<Partial<P>> {
+  return class WithNavigation extends React.Component<Partial<P>, State> {
     static displayName = `WithNavigation(${WrappedComponent.displayName ||
       WrappedComponent.name})`;
 
+    constructor(props: Partial<P>) {
+      super(props);
+      this.state = {
+        context: props.context || availableContext[0],
+        locale: 'en',
+      };
+    }
     handleContextChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-      this.setState({
-        context: e.target.value,
-      });
+      if ('jira' === e.target.value || 'confluence' === e.target.value) {
+        this.setState({
+          context: e.target.value,
+        });
+      }
     };
 
     handleLocaleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
       this.setState({
         locale: e.target.value,
       });
-    };
-
-    state: State = {
-      context: availableContext[0],
-      locale: 'en',
     };
 
     renderLocaleRadioGroup() {