+++ packages/media/smart-card/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @atlaskit/smart-card
 
+## 12.2.1
+
+### Patch Changes
+
+- [patch][b5eb352152](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/b5eb352152):
+
+  SL-336: fix page crash when state is undefined.
+
+## 12.2.0
+
+### Minor Changes
+
+- [minor][09f094a7a2](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/09f094a7a2):
+
+  SL-259: bump react-lazily-render, remove react-lazily-render-scroll-parent.
+
 ## 12.1.1
 
 ### Patch Changes